requirements.txt
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
discord.py==1.4.1
django-polymorphic==3.0.0
Django==3.1.1
idna==2.10
multidict==4.7.6
numpy==1.19.2
opencv-python==4.4.0.42
pytz==2020.1
sqlparse==0.3.1
typing-extensions==3.7.4.3
yarl==1.5.1